pyproject.toml
[tool.poetry]
name = "snaps-api"
version = "0.1.0"
description = ""
authors = ["ask4git <ask4git@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.16"
djangorestframework = "3.15.2"
gunicorn = "20.1.0"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.5.0"
django-allauth = "65.3.1"
dj-rest-auth = "^7.0.1"
requests = "^2.32.3"
cryptography = "^45.0.2"
rest-framework-simplejwt = "^0.0.2"
django-storages = "^1.14.6"
boto3 = "1.28.1"
django-cors-headers = "^4.7.0"
bson = "^0.5.10"
drf-spectacular = "^0.28.0"
botocore = "1.31.1"
pytest = "^8.3.5"
responses = "^0.25.7"
pytest-django = "^4.11.1"
drf-rw-serializers = "^1.4.0"
shortuuid = "^1.0.13"
firebase-admin = "6.9.0"
mock = "^5.2.0"


[tool.poetry.group.dev.dependencies]
coverage = "^7.9.2"
pytest = "^8.4.1"
pytest-django = "^4.11.1"
pytest-cov = "^6.2.1"
pytest-mock = "^3.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "snapsapi.config.settings.test"
python_files = ["test_*.py", "*_test.py"]
django_find_project = true
testpaths = ["snapsapi"]
